Move endpoint logic to handlers.go
handlers.go acts as a controller for http route definitions.
